+++ cmd/gossip/gossip/protocol.go
@@ -5,9 +5,9 @@ import (
 	"slices"
 	"time"
 
-	"github.com/antonionduarte/go-simple-protocol-runtime/cmd/gossip/membership"
-	"github.com/antonionduarte/go-simple-protocol-runtime/pkg/protorun"
-	"github.com/antonionduarte/go-simple-protocol-runtime/pkg/transport"
+	"github.com/antonionduarte/protorun"
+	"github.com/antonionduarte/protorun/cmd/gossip/membership"
+	"github.com/antonionduarte/protorun/transport"
 )
 
 // Protocol is the eager-push gossip protocol. Each node broadcasts a