+++ cmd/gossip/integration_test.go
@@ -9,10 +9,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/antonionduarte/go-simple-protocol-runtime/cmd/gossip/gossip"
-	"github.com/antonionduarte/go-simple-protocol-runtime/cmd/gossip/membership"
-	"github.com/antonionduarte/go-simple-protocol-runtime/pkg/protorun"
-	"github.com/antonionduarte/go-simple-protocol-runtime/pkg/transport"
+	"github.com/antonionduarte/protorun"
+	"github.com/antonionduarte/protorun/cmd/gossip/gossip"
+	"github.com/antonionduarte/protorun/cmd/gossip/membership"
+	"github.com/antonionduarte/protorun/transport"
 )
 
 const (